Try fixing autoFocus again, and remove all rAF usage (#16)
This attempts to fix the problem with `autoFocus` by breaking out of React's Effect ordering and instead immediately disabling existing layers whenever a new layer is _about_ to be added (i.e., it runs inline with the component code, not in a callback to be run after the tree has been committed). Since `autoFocus` happens when the tree gets committed to the DOM, this should ensure that existing layers will not intercept that focus event.